Converted pixel type to u32. Convert now filled with i32 casts to handle signed subtraction.

+++ vision/src/core/convert.rs
@@ -17,15 +17,20 @@ pub fn rgb_to_hsv(p: &mut Pixel) {
     let delta = v - cmp::min(cmp::min(p.h, p.s), p.v);
 
     let h =
-        if      v == p.h { ((255 * (p.s - p.v) / delta) % 1530) / 6 }
-        else if v == p.s { ((255 * (p.v - p.h) / delta) + 510 ) / 6 }
-        else if v == p.v { ((255 * (p.h - p.s) / delta) + 1020) / 6 }
-        else             { 0 };
+        if v == p.h {
+        	((255 * (p.s as i32 - p.v as i32) / delta as i32) % 1530) / 6
+        } else if v == p.s {
+        	((255 * (p.v as i32 - p.h as i32) / delta as i32) + 510 ) / 6
+        } else if v == p.v {
+        	((255 * (p.h as i32 - p.s as i32) / delta as i32) + 1020) / 6
+        } else { 
+        	0
+        };
     let s = delta / v;
 
-    p.h = h;
-    p.s = s;
-    p.v = v;
+    p.h = h as u32;
+    p.s = s as u32;
+    p.v = v as u32;
 }
 
 
@@ -34,9 +39,9 @@ pub fn rgb_to_hsv(p: &mut Pixel) {
 /// ## Parameters
 /// - p : input pixel to modify and convert to RGB
 pub fn hsv_to_rgb(p: &mut Pixel) {
-    let c = p.s * p.v / 255;
-    let x = c * (255 - (((6 * p.h) % 510) - 255).abs()) / 255;
-    let m = p.v - c;
+    let c = (p.s * p.v / 255) as i32;
+    let x = c as i32 * (255 - (((6 * p.h as i32) % 510) - 255).abs()) / 255;
+    let m = p.v as i32 - c as i32;
 
     let (r, g, b) = match p.h {
         0   ... 42  => (c, x, 0),
@@ -48,7 +53,7 @@ pub fn hsv_to_rgb(p: &mut Pixel) {
         _           => (0, 0, 0),
     };
     
-    p.h = r + m;
-    p.s = g + m;
-    p.v = b + m;
+    p.h = (r + m) as u32;
+    p.s = (g + m) as u32;
+    p.v = (b + m) as u32;
 }
